Add Rates class to represent the input and output event rates of the operators during the job.
Specify the InputRate class, which counts the number of events coming in the data stream and stores it in a HashMap

Create a Shedder class; a ProcessFunction to be used together with an operator. The Shedder will count events, storing then in an InputRate. It transmits the current InputRate every second to a side output stream to be collected by a coordinator.

Adjust minor details in the main job at DataStreamJob.