Add new classes for pattern detection with Finite State Machines.
It is now possible to declare patterns in a more straightforward way and to combine them together using the ComposeFSMProcessor class.

Documentation and integration with the normal operators will follow tomorrow.

Reorganise the classes in new folders and add smaller changes to the Measurement class.

Add Apache math commons in the dependency list.