Add new load shedding logic to the basic Analyser class, so that this class can be used, instead of the Coordinator, to calculate local shedding rates for their respective operator.
Improve the logging process to include the name of the operator or pattern in its output, so that we can trace events back to operators that created them.

Add new methods to EventPattern and Metrics to facilitate the calculation of the local load shedding rates.

Add minor adjustments in text and commentaries.