Add new layer of abstraction to the project. Jobs are now automatically parsed from the Settings using the new OperatorGraph. This classes automatically generates the flink-specific code for the query and connects all of its components accordingly. It defines the new standard of using the system.
Add new abstractions together with OperatorGraph to make it work. Operators are now observed together with all of their components in OperatorGroups (AbstractOperatorGroup, GlobalOperatorGroup and LocalOperatorGroup), this new classes group together the incoming and outgoing flink datastreams of an operator as well as its processing units (FSMOperator for global scopes and LocalOperator for local) and other components (in the global scope it includes the operator's Analyser and SourceCounter).

Add new abstraction for the processing logic inside the KeyedCo/ProcessFunctions used in the operators. Processing logic was generalised in the new class OperatorCore and specified in the new class GlobalOperatorCore and LocalOperatorCore for the 2 different processing scopes.

Add new Source class to define the system's sources in the OperatorGraph. Includes the source's information as well as the flink's datastreams.

Add new ways to determine the input and output types of an operator in OperatorInfo. One can now directly determine those either as a range or with an int array.

Add documentation to all new classes and improve older documentation.

Rearrange the project's structure and removes the old deprecated classes. Normal users should now only need to interact with the Settings class.

Fix minor bugs in the Coordinator's and OperatorCore logic as well as other small improvements.

This concludes the implementation of the first prototype.